refactor(input): consolidate remaining backends into src/input/ (#707 b)
Move sim-hid, pointer-service, and flutter-vm backends from src/tools/
into src/input/ for consistency with the #707a split. The old paths are
kept as thin re-export shims so all existing callers (tests, tools,
integration suites) continue to work without modification.

Internal callers that were already inside src/ are updated to import
directly from the new canonical paths:
  - src/input/backend-resolver.ts  → ./sim-hid-backend, ./pointer-service-backend, ./flutter-vm-backend
  - src/tools/diagnose.ts          → ../input/sim-hid-backend
  - src/tools/pasteboard-input.ts  → ../input/sim-hid-backend

src/metrics/input-telemetry.ts and src/metrics/input-telemetry-rollup.ts
already imported from ../input/backend directly (no change needed).

Behavior is strictly unchanged: no fallback semantics, no public API, no
tool schemas were altered.

Decision on pointer-service and flutter-vm: both moves are mechanical
(relocation + import-path update, zero logic change), so they are included
here per the task spec.
